Added logic to show when commits are staged
Originally was trying to use a regex OR, but the | character was being
interpreted by the shell and causing it to always return 0.  {}'s didn't
work either, so ended up relying on []'s and simple git status returning
an A or an M in the first column, as opposed to ?? or an M in the second
column.